chore: fix up navigation
