+++ frontend/package.json
@@ -15,7 +15,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
     "lucide-react": "^0.338.0",
-    "next": "^14.1.0",
+    "next": "^14.1.4",
     "react": "^18",
     "react-dom": "^18",
     "tailwind-merge": "^2.2.1",
