+++ frontend/src/app/confirm-reset/page.js
@@ -2,7 +2,7 @@
 
 import { Suspense } from 'react'
 
-import InvitationReset from '@/app/components/InvitationReset'
+import InvitationReset from '@/app/_components/InvitationReset'
 
 const Page = () => {
   return (
